README.md
This repository is for all the questions of linear data stuructures and non linear data structures .Here you will get all the code of each and every program in very efficient manner. You guys are the greatest asset who will help me for it. Lets crack each and every code.